Updates: Developing ingredientType JcomboBox and fixing previously selected item
Need to update:
If a NoneOfTheABove item already exists in the table and you try to change another ingredient to NoneOfTheABove, the table cannot go back to the original item.

Add Ingredients Type Column

MealPlan Ingredient Indexes

Repeated message on ingredient form with ingredient with no supplier info!

Change tableForm how data is stored to make when the ingredient info is changed easier
Meal Plan Needs to update with editForm