Updates: Refactoring Changing logic for updating macrotargets when an ingredient is added.
Need to Develop:
Efficient way of knowing if an ingredientName or type was changed as the whole ingredientsName doesn't need to be refreshed each time an ingredient is added.

refactor trigger columns

Repeated message on ingredient form with ingredient with no supplier info!

Change tableForm how data is stored to make when the ingredient info is changed easier